+++ examples/anim_util.c
@@ -22,7 +22,7 @@
 #include "webp/format_constants.h"
 #include "webp/decode.h"
 #include "webp/demux.h"
-#include "../imageio/example_util.h"
+#include "../imageio/imageio_util.h"
 
 #if defined(_MSC_VER) && _MSC_VER < 1900
 #define snprintf _snprintf
