+++ examples/vwebp.c
@@ -37,6 +37,7 @@
 #include "webp/demux.h"
 
 #include "../imageio/example_util.h"
+#include "../imageio/imageio_util.h"
 
 #if defined(_MSC_VER) && _MSC_VER < 1900
 #define snprintf _snprintf
