+++ extras/get_disto.c
@@ -23,8 +23,8 @@
 #include <string.h>
 
 #include "webp/encode.h"
-#include "../imageio/example_util.h"
 #include "../imageio/image_dec.h"
+#include "../imageio/imageio_util.h"
 
 static size_t ReadPicture(const char* const filename, WebPPicture* const pic,
                           int keep_alpha) {
