+++ imageio/Makefile.am
@@ -1,5 +1,5 @@
 AM_CPPFLAGS += -I$(top_builddir)/src -I$(top_srcdir)/src
-noinst_LTLIBRARIES = libexample_util.la libimagedec.la
+noinst_LTLIBRARIES = libexample_util.la libimageio_util.la libimagedec.la
 
 noinst_HEADERS =
 noinst_HEADERS += ../src/webp/decode.h
@@ -7,6 +7,8 @@ noinst_HEADERS += ../src/webp/types.h
 noinst_HEADERS += ../examples/stopwatch.h
 
 libexample_util_la_SOURCES = example_util.c example_util.h
+
+libimageio_util_la_SOURCES = imageio_util.c imageio_util.h
 
 libimagedec_la_SOURCES  = image_dec.c image_dec.h
 libimagedec_la_SOURCES += jpegdec.c jpegdec.h
